requirements.txt
langgraph
langgraph-supervisor
langchain-tavily
langchain[openai]
python-dotenv
braintrust
braintrust_langchain
rich
autoevals
modal
fastapi[standard]